requirements.txt
# Core data-science libraries
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.0

# Visualisation
matplotlib==3.8.4
seaborn==0.13.2

# Model persistence
joblib==1.4.2

# Configuration handling
pyyaml==6.0.1

# Progress bars
tqdm==4.66.4

# URL parsing helper (stdlib `urllib` is used, `tld` gives us top-level-domain info)
tldextract==5.1.2

# Optional dataset download (users authenticate with their own Kaggle API key)
kaggle==1.6.14

# Jupyter notebooks
jupyter==1.0.0
notebook==7.1.3
ipykernel==6.29.4

# Testing
pytest==8.1.1
pytest-cov==5.0.0
